Save correct AMR tet id of boundary face
This for later lookup of children for updating boundary faces and nodes
after refinement. Thanks to @rfbird for the help.

Also fixed and generalized how highwayhash was used to generate combined
hashes for node IDs for edges, faces, and tets.

Multiple initial uniform refinement steps now appear to be correct using
both CG and DG, tested with vortical flow.