Attempt to solve random deadlock with inciter
This commit replaces the two-step handshake-style collection of the
boundary nodes in Sorter by a single-step (still all-to-all style)
communication step. This appears to work better by (so far) not
triggering the quiescence callback used for debugging this (not
committed).

Potential caveat: The two-step algorithm was done that way because it
operated on a smaller data during an all-to-all type communication
pattern. It consisted a all-to-all broadcast + point-to-point answer
from all chares to all chares. Both of these steps operated on a list of
nodes potentially shared between two chares. The new single-step
algorithm, while simpler and appears to fix the problem, aggregates the
boundary node lists associated to chares for the whole problem on all
chares. This a larger data and a copy ends up on all chares (which then
promptly is thrown away after processing it). Since the data is larger,
it remains to be seen how this will scale and whether it may be a
bottleneck for scaling, but it appears that we no choice and have to do
it this way to fix the problem. Fingers crossed.