Generalize reading ExodusII side sets
This commit enables inciter to use ExodusII meshes that either contain
triangle element blocks or not.

Background
----------

For DG triangle face element connectivity information is required to set
boundary conditions. This is obtained from ExodusII side sets. Previous
to this commit, inciter has required the existence of triangle element
block(s) in the inciter input mesh. Triangle element block(s), however,
are not required to and therefore are not a prerequisite to
unambiguously define a mesh and corresponding side sets.

Accordingly, the ExodusII API allows reading side sets, independent of
whether there is a triangle element block in the file or not. If a
triangle element block is saved in the file, the triangle elements can
be used to define side sets but this is not required as side sets can
also be defined using a volume element + a relative face id within the
element adjacent to the side set. Thus, if triangle elements are not
saved in the file, side sets can still be specified by associating
volume (e.g., tetrahedron) element ids to side sets and a relative face
id which specifies which side of the volume element is adjacent to the
given side set. There is a convention defined in the ExodusII that
specifies the relative face ids for all volume elements supported, see
figure "Sideset side Numbering".

Changes in this commit
----------------------

To support meshes that do use triangle elements to define side sets as
well meshes that do not have triangle elements and define side sets via
tetrahedra only + relative face ids, this commit puts in code to the
ExodusII mesh reader that enables reading side sets either from triangle
or from tetrahedron element blocks.

Implementation details
----------------------

Note that there can be multiple triangle as well as tetrahedron element
blocks in arbitrary order in the file. (There can also be other cell
types but for now these are the two we care about and have code for.)
The ExodusII API to read side sets returns what is called "internal
element ids" + (element-relative) face ids (for tetrahedra: 1..4, for
triangles this is always 1). Internal element ids, as far as I can tell,
refer to a file-internal ordering, i.e., the element ids in the order as
the element blocks are numbered in increasing order in the file. Since
we read potentially multiple triangle and tetrahedron element blocks
independently, i.e., all triangle blocks and all tetrahedron element
blocks separately and store their connectivity as two independent arrays
(one for all triangles and one for all tetrahedra), the internal element
ids must be converted to triangle element ids and tetrahedron element
ids while also identifying which cell type a given side set internal
element id refers to. With that information the triangle face
connectivity can be correctly extracted from the independently read
triangle and tetrahedron connectivities.

Thus this approach is more general compared what we had before as it
allows using meshes that do have triangle element blocks (which may be
used to define side sets and thus their face connectivity) as well as
meshes that only define tetrahedra and use those to define side sets and
their face connectivity.

Current status
--------------

This commit appears to be correctly generating the side face
connectivities from both types of meshes in serial only. More testing
and regression tests exercising both types of meshes are of course
needed. But next is to figure out how to find the face triangle element
connectivities also in parallel, where not all PEs have access to the
full tetrahedra and/or triangle element blocks but only a portion of
them.