Remove those face/node lists at which no BCs are set
This commit removes those face/node lists, read from file, associated to
side sets, to which the user did not configure BCs on. Those will not be
needed during solution.

This not only enables not carrying boundary face/node lists further on
during initialization, but also ensures that input meshes in inciter
that define multiple side sets (potentially overlapping faces with other
side sets) that, if all augmented, do not necessary comprise a non-leaky
boundary, as measured by DG::leakyAdjacency(), do not trigger an error.

In other words, a problem's boundary conditions, specified in the input
file for the given problem and mesh, are now solely responsible for
defining a non-leaky boundary.

This has been tested with meshes which do contain multiple side sets of
which a subset together (2...7) define a correct non-leaky adjacency,
but also contain an extra side set (1) which contains all side sets in
2...7.