Move unittest definitions to the unittest executable
This way when the unit tests register themselves with the global
singleton runner, defined in UnitTest.C, using an external reference to
the runner in the tut include, at link time, all objects that reference
the runner will be visible to the linker in a single step, when linking
the unittest executable.

Previously the tests, now appearing in multiple compilation units were
bundled up into a library which in turn was linked in when creating the
executable. While this two-step rpocess worked fine with dynamic
linking, static linking, using an intermediate static library, the
linker failed to find the same global runner between the executable link
and the library.