Reduce boundary node lists for side sets to PE only
This is a follow up to 6d02894f. Now the node lists associated to side
sets are stripped in Partitioner to those that exist on the given PE
only. This is what is passed to Refiner, which then will have less work
to do, since when due to AMR side set data is updated, only faces and
nodes on side sets that exist on a given PE will be searched. Caveat:
not yet tested with AMR.