Add chare state collector to all executables
This commit adds a Charm++ chare state collector to all excecutables.
This is intended for debugging and thus it is off by default, but
automatically on when quiescence detection is enabled.

Chare state collection and screen output can be enabled by the -S or the
--state command line arguments. The chare state is displayed after a run
is finished and the data collected is grouped by chare id (thisIndex),
and within groups data is ordered by the time-stamp when a given chare
member function is called. See src/Base/ChareState.h for details on what
is collected.  Note that to collect chare state, the given chare must be
instrumented. For an example, see src/Inciter/DG.h. Note that if
quescence detection is enabled, chare state collection is also
automatically enabled, but the chare state is only output if quiescence
is detected (which also triggers an error).

The data collected and the way it is displayed may change in the future.