Update regtest baselines that now have nodesets
Since we now save node sets if a nodal discretization scheme was used,
regression test result exodus files now also contain node sets there
were used to specifiy boundary conditions. This is not really an error,
but was not discoeverd so far because exodiff output was silenced in
cmake/ctest. Since cmake@81b4a4a now enables the error output, this is
now caught and so this commit updates the baselines for those tests.