Switch Partitioner to Charm++ nodegroup
This commit switches Partitioner from group to nodegroup. This means that
instances of Partitioner are now created by the runtime system one per compute
node and not necessarily one per PE as before. This behavior is thus different
only with an SMP-build of Charm++ and with using +ppn N with N>1.

As a result Partitioner now reads the mesh using a reader per compute node
instead of a reader per PE. This potentially gives a much better performance on
large core (and compute node) counts, i.e., for large problems.

Caveats:

(1) Only DiagCG and DG works for now, MatCG does not work with +ppn N, where N
> 1. Note that none of the Inciter/MatCG regression tests specify +ppn and the
default is 1, so this will not break regression tests and the since the
behavior of nodegroups are the same as groups in non-SMP mode, nothing should
change in non-SMP mode.

(2) While Partitioner reads the mesh a worker per compute node and that's most
likely beneficial for large problems, everything else that follows the mesh
read in Partitioner, i.e., mesh partitioning with Zoltan2 and mesh
re-distribution to compute nodes based on the result of the partitioning now
happens only a single worker per compute node. This is not good because we
loose parallelism for these tasks. A better solution here would be to separate
the mesh read into a nodegroup and everything else that follows to a group.

(3) SMP mode with +ppn N, N>1 has no regression tests yet.

(4) Unittest hangs even with +ppn 1, within the MPI suite.

(5) Running the regression test suite in SMP mode is fine, but doing
nothing (or leaving our current setting) on the thread placement yields
massive thread oversubscription currently.

In summary, SMP mode is experimental and WIP at this point.

More in this commit:

 * cleanup of unused member data in Partitioner
 * remove screen output on misleading linear load distribution
 * extend screen output on work units to include distribution of the problem to
   compute nodes and PEs
 * handle error discovered with Zoltan2's mj partitioner that in some
   combinations of meshes, compute nodes, and worker threads (=N in +ppn N)
   leads to no mesh elements assigned to a compute node; this might go away
   when we, again call Zoltan2 from a group
 * Solver::created now appears to be unnecessary, yay!