Do not specify PE for dynamic array insertion in inciter
In SMP mode this allows the runtime system to better distribute chare
array elements across the available threads/PEs, and thus yields orders
of magnitude speedup in SMP mode, which is now faster than non-SMP mode,
which what we expect. Tested on a 2-socket shared-memory machine with
two dedicated communication threads, one for each socked and explicit
mapping of threads, e.g.,

./charmrun +p 36 Main/inciter -v -c ../../tmp/vortical_flow_dg.q -i ../../tmp/unitcube_60k.exo -b +ppn 18 +pemap 1-17,19-35 +commap 0,18

which configures 2 compute nodes each with 17 worker threads and a
designated communication thread.

Huge thanks to @ebohm for the ongoing help with this.