Remove MPI-interop from unittest
This commit revamps unittest to invoke MPI unit tests from a Charm++
nodegroup instead of using MPI-Charm++ interoperation. This considerably
simplifies the harness, because there is no need to do a separate MPI
harness after the serial and Charm++ ones. This also is a first step
towards porting unittest to Charm++'s SMP mode, which now works but only
for +ppn 1. There are two items left for this to be fully ported:

1. Fix the MPI unit test "ErrChkMPI macro doesn't throw all true", which
currently fails.

2. Port those unit tests that formerly were not part of the MPI suite
but still invoke MPI calls. These are MPI-only libraries, such as
ExodusII and Omega_H.