Working AMR for multi stage refinement
The fix here revolves around how the compatibility works, where by a 1:8
neighboring a 1:4 and sharing an intermediate node only the 1:4 has
compatability class 3. 1:8 has compatability class 2 (locked but no
intermediate) despite sharing a node.

It also overhauls how the book keeping was done to track intermediate
and locked edges and fully regenerates a set which (I think) removes a
subtle race condition from the way I was processing it dynamically
before

This currently has a *very* expensive over kill fix in place, which
loops over all tets at the end of the step. I need to make it more
sensible and have a plan for this..