Convert all-to-all style bnd node comm to p2p
This commit changes the all-to-all style setting up communication maps,
started in Sorter's constructor to point-to-point style using a
distributed table. It also removes two all-to-all comm steps that used
to snowball a lot of boundary node list data across chares and replaces
it with communicating this data during reshuffling that's already done
after mesh partitioning by Partitioner.

This leads a lot better scalability of this setup step, especially at
large scales and for large meshes. Tested, using Charm++ Projections on
10 compute nodes (36 cores/node) using a 18M-cell mesh with DG. DiagCG
also works, but [CI SKIP] because MatCG is not yet ready for SMP.

This is part of the work with, and thus many thanks to, the Charmworks
folks who are a great help in pointing out and explaining potential
causes of such problems and teaching us how to use their Projections
tool.