Add inciter's parser understand 'scheme dgp1'
This is besides the existing 'scheme dg', which now means DG(P0).

Also put in coding to query the centering for a scheme, which
returns NODE for DiagCG and MatCG and ELEM for DG. This makes client
code simpler when the centering and not the specific discretization type
is in question.