Quiet srd::cerr in unit tests
This commit adds the capability to quiet std::cerr and uses it to
redirect its output for unit tests that test throwing exceptions to a
string stream which then is thrown away as it goes out of scope, which
effectively quiets std::cerr's output temporarily (within the scope cerr
was quieted).

This is necessary from unit tests to not output garbage to the screen
because throwing an exception now outputs its message, passed in at the
exception call-site from Exception's constructor. In the past this
message has been output from Exception's handleException() member
function but now we output it from the constructor. This in turn is
necessary because some exceptions, thrown from asynchronous code, are
only caught by the terminate handler, in which the original exception
message is not available because no arguments can be passed to the
terminate handler. In the terminate handler we re-throw another
exception to generate a call trace but it is only at the original
exception call site where we have a good error message and
file:line:func info on the error/exception.

Long story short: Exception now outputs its message, just passed in to
its ctor, at its call site, to std::cerr to output a meaningful
informative error message, which are now silenced for unit tests that
are explicitly designed to test exceptions.