Modify logic to incorporate t>0 AMR
This commit puts in stubs and modifies the end-of-time-step logic in all
of DiagCG, MatCG, and DG to call the mesh refiner object during time
stepping. Refinement is not yet called from any of the discretizations,
but Refiner's end-of-refinement-step logic is modified to incorporate
t>0 AMR and augmented with a call to update the mesh in Discretization
(which holds the mesh) after a refinement/de-refinement step during time
stepping.