dtref appears to work with DiagCG in parallel
This commit adds the capability to DiagCG to perform AMR during time stepping
(dtref). Both serial and parallel appear to work. Caveats:

 - no BCs,
 - uniform refinement only,
 - severe reallocation of memory for everything mesh-related, i.e.,
   unknown/solution arrays, so expect this to be quite slow for now,
 - will only for refinement, i.e., adding new data and not for de-refinement,
 - solution values at newly added nodes are simply arithmetic averages of
   edge-end-point values, which will not fly well with de-refinement.

Details:

* The default filename in which inciters saves its field output has been
changed. This is to follow the convention described at

https://www.paraview.org/Wiki/Restarted_Simulation_Readers

and also documented in Discretization::filename(). This is necessary to enable
a changing mesh for exodus, which is originally not designed for this purpose.
With such naming, however, enables paraview to stitch together a simulation
output that may contain different-topology meshes.

In a nutshell, the output is now named as, e.g., out.e-s.0.16.0, where
 - e-s denotes ExodusII sequence
 - the first number counts the number of times the mesh topology changes,
 - the second number is a constant and equals the number of chares,
 - the third number is the chare id.

Each of such files may contain multiple time steps. Accordingly, all inciter
regression tests had to be renamed. This works great, because now runs in the
same directory using different number of chares doe not clobber each other -
slightly easier to work with.

* Updated how tk::keyEqual() compares the keys of the two containers. Though
not used for what I changed this for, I keep this this way because this can
test unordered containers and the previous implementation was looping through
the two containers in lock-step so it only worked for ordered containers.

* Add new keyword, dtfreq, which configures the refinement frequency, with
basic error checking during parsing. The default is 3 (every 3rd time step).

* The end-of-time step logic had to be changed to accommodate for the correct
resizing of various data structures after a refinement step. This affects both
the worker (children, MatCG, DiagCG, and DG) as well as Discretization and
Transporter, due to added communication that are needed for re-computing the
left hand side and recomputing the mesh volumes (stored in various ways in
Discretization). Both of these are overlapped asynchronously though.

* No attempt has been done to exercise during-time stepping refinement for
MatCG and DG at this time, but they pass their regression tests, none of which
exercises dtref.

* Code reuse is quite nicely possible, especially in refiner, but also in
recomputing and resizing data structures, everywhere needed.