Keep DGPDE::analyticSolution but cleanup and rename
On a second thought, I do believe having a member function that only
queries the analytic solution from a PDE makes sense, of course, only if
a PDE, as configured by its Problem policy class, defines it.

Separating the ability to simply only query the analytic solution,
independent of how it is used, yields a cleaner code, compared to the
previously used piggybacking on initialize() and using that for two
purposes: (1) initializing (a potentially large amount of data at once)
and (2) querying the analytical solution.

The benefits are cleaner and separate/independent APIs for initialize()
and analyticSolution(), and potentially better performance when the
latter is needed, e.g., in ElemDiagnostics::computeP0, which now does
not call initialize() for a large array only for a small local data.