Add MixDirichlet SDE to walker
A derivative of the Dirichlet SDE, which will be constrained to
multi-material mixing in tubrulent flow. For now, compared to Dirichlet,
it computes the additional quantity, particle density and correctly
reproduces the unstrontrained Dirichlet SDE behavior. Next is to
implement Ray's two new specifications for S_alpha and test if d<rho>/dt
= const.