Switch AMR correction iteration to before refinement step
This commit changes the logic of how correction iterations are done,
required after a non-uniform refinement step in parallel to arrive at a
conforming mesh.

Instead of doing a refinement and updating the mesh then doing
correction iterations, we now only tag edges to be refined, perform
potentially multiple correction iterations, involving communication
along chare boundaries, until the iterations converge to tagged edges on
the boundary on which all chares agree, then perform mesh refinement and
mesh update.

Compared to how we did this previously, this yields the same mesh in
serial and in parallel and the resulting mesh appears to be independent
of the number of chares.