Remove default runner args used for regression tests by ctest
This commit removes the default parallel job runner arguments, which has
been "--bind-to none -oversubscribe". This should really not be
hard-coded, especially that these do not work with older openmpi
versions.

Instead, the user can specify the RUNNER, RUNNER_NCPUS_ARG, and
RUNNER_ARGS to fully customize the runner. If no runner is given, the
default is the same as before: charmrun. But this now allows, e.g., to
specify mpirun as the runner, e.g.,

cmake -DRUNNER=mpirun -DRUNNER_NCPUS_ARG=-n -DRUNNER_ARGS="--bind-to none -oversubscribe",

which also shows how to pass extra optional arguments to the runner.
This is also handy, when charmrun breaks for whatever reason, e.g.,
https://lists.cs.illinois.edu/lists/arc/charm/2018-10/msg00010.html.