AMR lock iteration now communicates THEN collects edges to still be corrected
In the AMR PE-boundary iteration used to correct edges, so far when we
returned to Refiner for a new iteration, we started by looking for
extras again without first communicating the edge data. Now I changed
this so that a single iteration consists of two steps: (1) communicate
edges, (2) look for extra edges (and correct, i.e., call the
compatibility algorithm).

Another minor change is that the exit condition for parallel
communication of chare-boundary edges is now slightly different:
completion is now detected by checking the size of m_ch on the receiving
side instead of testing for the same on the send side after receiving a
receipt. This new one is simpler, less code, and potentially more
performant because m_ch is a symmetric map on all chares. (The previous
solution did not rely on the symmetry.) As a result recvRefBndEdges() is
now gone.

Also switched the data type for edge data from tuple to pair.

As before, still produces a non-conforming mesh along chare-boundaries
after a 2nd non-uniform refinement in parallel.