Pass a nonzero exit code on error to Charm++'s CkExit
This has not been possible before Charm++ v6.9.0. Previously on an error
we called CkAbort() to exit with a nonzero exit code which was
hard-coded to CkAbort(). Now we call CkExit() and pass a nonzero exit
code. The advantage is that mangled output from Charm++ does not go to
the screen on an error. (We already generate our de-mangled trace, so
Charm++'s is not necessary.)