Walker: joint correlated Gaussian ICs
Walker now can initialize any SDE system with a joint PDF of arbitrary
dimensions with a prescribed mean vector and covariance matrix.

Note that of the three RNG libraries interfaced, only MKL provides a
library function that can sample from a joint correlated multi-variate
Gaussian.

Some error checking has also been put in for walker's parser that
catches if the user selects the joint correlated Gaussian initpolicy but
does not configure a MKL RNG for the equation, or if the user does not
specify the mean and/or covariance matrix. No error checking is done on
the covariance matrix being positive definite.

Added unit tests to test MKL's MCG31 and MCG59 used for sampling from
multi-variate Gaussians using 3x3 and 5x5 covariance matrices.

Also tested this with integrating the Dirichlet SDE starting from a 2x2
joint Gaussian and evolve in time, which is surprisingly robust as the
Dirichlet SDE is not even defined for samples outside of zero and unity.
Nevertheless, this test verifies that the Dirichlet SDE can be used to
evolve initially positive covariances towards zero, even though its
stationary state only allows negative covariances, see also
https://www.hindawi.com/journals/ijsa/2013/842981/.