Switch calls to Solver's member functions to direct ones
This basically reverts many of the member function calls, issued from
MatCG to Solver, from via operator[ node() ] to ckLocalBranch(). This is
how this was previously, but for some reason I don't recall, porting
Solver to nodegroup appeared to require these calls via the former.
However, that way creates random dead-locks (not quiescence). Will still
have to see how this behaves in SMP mode, but for now this switch (back)
appears to fix the random dead-locks with MatCG.

Also instrument MatCG for chare state collection, but leave commented
for now.