Convert Solver::*import data from map to vectors
This is a first step in an attempt to fix a race condition that is now
possible (and potentialyl happening) with Solver as a nodegroup. My
thinking is that member functions, e.g., charerow(), can now be called
by the runtime system concurrently on the same Solver nodegroup object
instance, which then can create a race condition as it writes to
m_rowimport at the same time from different calls. The solution this
commit puts in is to use a vector of vectors instead of a map, whose
write locations are pre-allocated. The size of the outer vector equals
to the total number of chares across the whole problem on all Solver
nodegroups, but only those are used that end up with a non-empty inner
vector. To make this solution work, other member data, such as m_row,
will also have to be converted.