Make quieting std::cerr thread-safe
This commit changes how redirecting of std::err is done in unittest.
This is needed so messages do not come to the screen (std::err) during
throwing exceptions.

Quieting std::cerr is done by redirecting its stream buffer into a
std::stringstream and then resetting its state when quieting is no
longer needed.

Previously, this was done locally, immediately before and after the
exception was thrown at every location in a unit test. In SMP mode,
however, this is not thread-safe as std::cerr is a global, not a
thread-local, resource, i.e., shared among multiple threads. As a
consequence, multiple unit tests executing concurrently would step on
each other's temporary streams as change std::err's state.

With this commit, we redirect std::cerr in a Charm++ nodegroup's
initnode routine, which is guaranteed to be called once per logical
node, per which there appears to be a std::cerr. The new Charm++
nodegroup is called QuietCerr and is under src/UnitTest.