Fix executables return code
This fixes a bug in how return codes to shell were generated.
Previously, if quiescence was enabled, executables incorrectly did not
set a nonzero exit code when, e.g., a test failed or threw an exception
un unittest.

With this commit returning the exit code to the shell is now independent
of whether quiescence was enabled or not.

Note that this affect every executable, since tk::Init is generic.