Make MPIRunner::rungroup exclusive in unittest
This makes the Charm++ nodegroup, MPIRunner, exclusive, which ensures
that calls to this entry method do not run concurrently on a single
logical node. As a result test groups that need to be executed in a
thread-safe manner, i.e., serialized with respect to each other, are
serialized. (The rest of the thests are still independently spawn.)