Move two DerivedData unit tests to their own group
The new groupname is DerivedData_MPISingle, which ensures that tests in
that group are serialized as they run by the unit test harness with
respect to tests in other test groups that have the string "MPISingle"
in their groupname.

These two DerivedData tests use ExodusIIMeshReader, which calls HDF5,
which is not thread-safe by default (which is how we use it).