Add new PDE type 'MultiMatCompFlow'
This commit adds a new PDE type that will be used for multi-material
flow algorithms. It adds keywords, parsing, etc., and it is similar to
the existing 'CompFlow' PDE type. One main difference is that
'MultiMatCompFlow' needs the number of materials, which can be set by
the user via the 'nmat' keyword. Example:

inciter

  scheme dg

  multimat_compflow

    depvar c
    physics veleq
    nmat 2
    problem vortical_flow

    ...

  end

  ...

end

So far there is a single physics configuration available under
multimat_compflow, 'veleq' which stands for velocity eqilibrium.