walker: add const-dissipation velocity policy
This commit enables walker to

* compute a forced homogeneous turbulence using the existing (simplified
or generalized) Langevin models via the new policy const_dissipation,

* impose a mean scalar gradient in the mixmassfracbeta mix model via
coupling to the instantaneous velocity field.

It also removes the mandatory requirement of a dissipation coupling from
the velocity model, so that a dissipation equation does not have to be
carried if a constant dissipation policy is used for the velocity
equation.

Some cleanup on the dissipation, velocity, and mixmassfractionbeta SDE
implementations.