Send refinement data on all edges
Instead of sending refinement data on the chare-boundary edges only, we
now send this data on all edges of a chare. The communication pattern is
still point-to-point but we now send more data. This is to simplify
debugging. As a result, Refiner::m_bndEdges though communication, now
is NOT used when doing the iterative correction logic on shared edges.
This is probablly inefficient, but this definitely ends up checking more
edges than before.