AMR: edgelist initial refinement is now treated as other initial refinement types
Instead of an add-on to the existing initial refinement list.

This leads to simpler code that is more logically concistent with how
other initial refinement types are handled and also allows
edgelist-refinement to be combined in arbtirary fashion with other
initial refinement types, i.e., use it as a 2nd step not only the 1st.