Remove reducing npoin
This commit removes the global sum of the number of mesh points, both
before and after initial refinement, because its final result is
misleading: the sum of the number of mesh nodes in parallel is more than
the number of nodes. The number of cells still indicates the correct
number of cells across the whole problem.