Store relative id of coupled eq instead of offset only
This commit replaces the way information on a coupled eq is stored in
walker between DiffEqs. Previously, if an eq was coupled to another,
only the equation system offset was stored so that the instantaneous
variables were directly accessible from the equation coupled. This commit
computes and stores the relative equation index of the coupled equation,
associated to the equation. This allows accessing the full
configuration of the coupled equation, so that, e.g., the mix equation
can query if the velocity equation solves for the full or the
fluctuating variable, etc.

Also, the constant dissipation coefficients policy is now replaced with
the stationary policy, which thus can now be used to denote, in general,
a policy that is used to force the PDF constant in time. See
DissipationCoeffStationary and Coefficients::src() in
DiffEq/Dissipation.h and VelocityCoeffStationary() in
DiffEq/VelocityCoeffPolicy.h.

Also, functionality to query information on a DiffEq configuration,
e.g., for screen output, on coupled equations is now abstracted away to
generic functions in DiffEq/CoupledEq.h. DiffEq/CoupledEq.h also
contains accessors for a number of items of the configuration of coupled
equations, e.g., offset. This makes client code related to coupled
equations shorter, a lot more uniform, and thus less error-prone.

Also add new MixMassFracbeta coefficients policy,
MixMassFracBetaCoeffInstVel, which enables coupling the beta mix model
to be coupled to the instantaneous velocity model, e.g., the Langevin
equation, and extract and use its statistics and/or instantaneous
velocity.