Add exodiff map option for dgp2 serial tests
The way we massage our data structure for physical boundary element
connectivity, as it starts from reading it from the exodus file, passing
through mesh partitioning, reordering, etc, until it arrives at DG.C
(and is written out to file in Discretization::writeMesh), is not
necessarily deterministic. This is the case even on a single CPU, e.g.,
due to different compilers and/or compiler options, because, e.g., we
use maps and unordered maps. Thus the order in which the faces are
listed in the values of the map of vectors, e.g., Sorter::m_bface is not
always the same. However, the connectivity must still be same for each
side set, i.e., it must contain the same number of triangles and the
same connectivities for each face, only in potentially different order.

The fix is to pass -m to exodiff, which is almost always required in
parallel:

From exodiff --help:

 -map (Invokes a matching algorithm to create a mapping between the
       nodes and elements of the two files. The topology must still be
       the same (within tolerance), but can be ordered differently.)

The symptom becomes obvious if one diffs the ncdumps on the std and the
out files, e.g.,

ncdump nleg_dgp2.std.exo > s
ncdump out.e-s.0.1.0 > o
vimdiff s o

which shows that the files diff in their block 2 (triangle faces), and
the connectivities are there only their order is different.