Inciter in SMP: Serialize field output per compute node
This commit adds a new class in IO/MeshWriter.[Ch] and IO/meshwriter.ci.
MeshWriter is a Charm++ group (i.e., guaranteed to be a single instance
per PE whose instances never migrates with Charm++'s automatic load
balancing) that facilitates serializing field output per compute node.

This is required in Charm++'s SMP mode and (as it is implemented
currently) should not make any difference in non-SMP mode. The way field
output happens is not changed: we still write as many files as the total
number of chares.

MeshWriter is class Charm++ group because we use Charm++'s guarantee of
invoking an entry method of a group one at a time to serialize
potentially multiple writes from the same PE.

In non-SMP mode the difference compared to previous behavior is the
serialization of multiple files for non-zero overdecomposition, so this
is potentially better performance, although this has not been measured
at this point.

In SMP mode the serialization channels through all writes via the first
PE of every compute node. This is also potentially better performance,
especially at large core counts, where the I/O concurrency of hardware
is lower than the compute concurrency (which is usually the case for
large core counts), but this has also not been benchmarked at this
point. On the other hand, channeling writing of multiple files via a
single PE of every compute node is required for NetCDF, HDF5, and
ExodusII, the libraries we use for field output, as none of these
libraries are thread-safe and miserably fail for +ppn > 1, i.e., a
runtime setting that would exercise SMP mode.

Other minor details:

The enum Centering has now been moved from
Control/Inciter/Options/Scheme.h to Mesh/Centering.h, so that we can
also use it in MeshWriter and not only in Scheme.

The details of element-centered vs node-centered field output as well as
the file type (ROOT vs ExodusII) are now hidden behind MeshWriter. This
cleans up client code (Discretization and its children).

At this point all of MeshWriter's entry methods have the 'sync'
attribute, which means they block until they finish. This is simply a
punt to reproduce the previous behavior of writing field output (and
mesh) into files serialized from multiple chares. A potentially better
way of doing this is remove the blocking and make these entry methods
asynchronous to overlap the writes with computation, which is a task for
the future.

TODO:
* The regression tests with migration still fail.
* Add more npe != ppn regression tests for SMP mode.
* Augment CI with SMP mode.