Mods to fix migration tests in non-SMP mode
With this commit all tests pass in non-SMP. We have extra tests in SMP
mode to excercise SMP mode with npe != ppn but more needed. There is
also more work needed for the mesh writer for SMP mode and to pass the
tests in SMP mode. See progress and more details at
https://github.com/quinoacomputing/quinoa/issues/261.