Apply changes required for testing migration
This commit is applied on top of the last known state of the work
towards migration with smp-compatible field output. It is based on a
cherry pick but is more that that.

Details:

Migration is non-blocking.

Disable field output for now.

Re-enable quiescence detection for migration tests, which works fine
with non-blocking migration. It still detects quiescence randomly but I
suspect this has been there for a long time and does not appear to be
related to the random segfaults and deadlocks associated with the
mixture of bound arrays, migration, and sync/threaded entry methods.

Tested with the following command unitl break:

while ctest -j36 -R asynclogic -L migration --output-on-failure; do :; done | c++filt

Only random quiescence is triggered, no random segfaults, only with the
gnu compler random deadlocks. I do not assume these last two problems
are associated with the random segfaults and deadlocks seen recently.