Move initial wait4lhs to ALECG and DiagCG ctor
The earlier the better to make this call to avoid these DAGs potentially
not yet enabled but the class receiving a message to comlhs(),
potentially completing all receiving the lhs contributions and calling
comlhs_complete without the DAG being enabled.

This does not appear to make any difference butit appears safer to make
this call in the ctors.