Switch the order of 'dtref' and 'dtref_unifor' in grammar
This is pretty sublte has come up before. If keyword a is a subset of b
in a grammar rule in tk::grm::block< a, b, ... >, b will never be
matched only a. However, we the grammar is switched so that the longer
keyword comes first followed by the shorter one, both matches correctly.

This order in the grammar should not matter and should be fixed. This
commit only fixes the grammar for 'dtref' and 'dtref_uniform' so that
dtref_uniform is also matched and not only dtref.