Update reordered mesh in Refiner from Sorter
This commit puts in code to query, receive, update the mesh, held by
Refiner, if/after Sorter has performed PE-locality mesh reordering.

This is required because Refiner is created before Sorter and if Sorter
has performed a reordering, based on user input, which is optional, the
mesh that Refiner was initialized with is then different.

Refiner passes a callback to one of its member functions,
Refiner::reorder, a Charm++ entry method, when it creates Sorter, which
then Sorter calls if (and only if) it performed a reordering.
Refiner::reorder then issues a direct, i.e., non-entry-method, call to
Sorter to query and receive the reordered mesh and updates its data
structures, so that dtref can be done on the correctly ordered mesh that
is consistent with what the workers, e.g., DiagCG, have.

Refiner also have to update its AMR object with the new mesh and
currently this is done by re-creating it via its constructor and a copy
assignment, which then overwrites the original AMR object in Refiner.
This seems to be okay for now, as this seems to correctly update the
mesh but is almost certainly not what we ultimately want because this
deletes the AMR lib's history of t0ref. A better way to do this will
probably be to expose an API which can be used to update the mesh in the
AMR lib object.

Since we now need to call a copy constructor on the AMR lib, I removed
some const qualifiers from mesh_adapter_t and refinement_t.

Also renamed belem to bface in Refiner to make it consistent with other
classes, e.g., Sorter.

Also rename Refiner::t0refstart to Refiner::start because that better
expresses the fact that that is the start of a refinement step,
independent of it is called during t0ref or dtref.

Also removed the m_initial-logic from Refiner::t0ref because that is now
only called during t0ref.

Also add two new dtref regression test, exercising mesh reordering with
t>0 refinement.